# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RRHH API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/puesto-grupos": {
            "get": {
                "tags": [
                    "Puesto Grupos"
                ],
                "summary": "getPuestoGrupoList",
                "description": "Get all PuestoGrupos",
                "operationId": "c7e9e6b8763bf1f6381181e7ab80e9e3",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Puesto Grupos"
                ],
                "summary": "createPuestoGrupo",
                "description": "Create PuestoGrupo",
                "operationId": "c63ce87c486097c560fbb4e2f9dc9ad7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/puesto-grupos/{id}": {
            "get": {
                "tags": [
                    "Puesto Grupos"
                ],
                "summary": "getPuestoGrupoItem",
                "description": "Get PuestoGrupo",
                "operationId": "7441a688867b42c09a29d3912c4d0933",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PuestoGrupo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Puesto Grupos"
                ],
                "summary": "updatePuestoGrupo",
                "description": "Update PuestoGrupo",
                "operationId": "9939f5b04258e8510eacac0e897a1950",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PuestoGrupo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Puesto Grupos"
                ],
                "summary": "deletePuestoGrupo",
                "description": "Delete PuestoGrupo",
                "operationId": "9abcae48207a0e4b84a42c9212421c31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PuestoGrupo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    }
}